More separation of concerns In-memory models cache Cleanup
