+++ app/src/main/java/com/netchar/nicknamer/domen/service/NicknameGeneratorServiceImpl.kt
@@ -21,7 +21,7 @@ import com.netchar.nicknamer.domen.models.Nickname
 import com.netchar.nicknamer.domen.models.Nickname.Companion.orEmpty
 
 class NicknameGeneratorServiceImpl(
-        private val dataSource: NicknameRepository
+        private val repository: NicknameRepository
 ) : NicknameGeneratorService {
     companion object {
         private const val PEOPLE_PREFIX = "people"
@@ -29,23 +29,23 @@ class NicknameGeneratorServiceImpl(
 
     override suspend fun generateNickname(config: NicknameGeneratorService.Config): Nickname {
         val key = "${PEOPLE_PREFIX}_${config.gender.value}_${config.alphabet.value}"
-        val source = dataSource.getModels()
-        return source[key]?.generateNickname(config.nicknameLength).orEmpty()
+        val models = repository.getModels()
+        return models[key]?.generateNickname(config.nicknameLength).orEmpty()
     }
 
     override fun addToFavorites(nickname: Nickname) {
-        dataSource.addToFavorites(nickname)
+        repository.addToFavorites(nickname)
     }
 
     override fun removeFromFavorites(nickname: Nickname) {
-        dataSource.removeFromFavorites(nickname)
+        repository.removeFromFavorites(nickname)
     }
 
     override fun getFavoriteNicknames(): List<Nickname> {
-        return dataSource.getFavoriteNicknames()
+        return repository.getFavoriteNicknames()
     }
 
     override fun isFavorite(nickname: Nickname): Boolean {
-        return dataSource.isFavorite(nickname)
+        return repository.isFavorite(nickname)
     }
 }
