+++ app/src/main/java/com/netchar/nicknamer/presentation/ui/about/ContactsAdapter.kt
@@ -17,7 +17,10 @@
 package com.netchar.nicknamer.presentation.ui.about
 
 import android.view.ViewGroup
+import androidx.annotation.DrawableRes
+import androidx.annotation.StringRes
 import androidx.recyclerview.widget.ListAdapter
+import com.netchar.nicknamer.R
 import com.netchar.nicknamer.databinding.RowContactBinding
 import com.netchar.nicknamer.presentation.infrastructure.helpers.BindableViewHolder
 import com.netchar.nicknamer.presentation.infrastructure.helpers.DefaultDiffCallback
@@ -44,4 +47,10 @@ class ContactsAdapter(private val listener: (Contact) -> Unit) : ListAdapter<Con
             override fun from(parent: ViewGroup) = ContactViewHolder(RowContactBinding.inflate(parent.inflater(), parent, false))
         }
     }
+}
+
+sealed class Contact(@DrawableRes val image: Int, @StringRes val description: Int) {
+    object Instagram : Contact(R.drawable.ic_instagram, R.string.about_label_follow_on_instagram)
+    object LinkedIn : Contact(R.drawable.ic_linkedin, R.string.about_label_connect_on_linked_id)
+    object Mail : Contact(R.drawable.ic_gmail, R.string.about_label_send_email)
 }
